providers.py: Add PREFERRED_RPROVIDER
[Yocto Bug #6149] - https://bugzilla.yoctoproject.org/show_bug.cgi?id=6149

Use PREFERRED_RPROVIDER to specify runtime packages in the case of multiple
runtime providers.

The WARNING from bitbake suggested specifying a PREFERRED_PROVIDER for the
package name, but setting it did nothing because the filterRunTimeProviders
searched only for the recipe names.

For the example case in bug #6149, bitbake looks for PREFERRED_PROVIDER_openssh
and PREFERRED_PROVIDER_dropbear. This change allows
PREFERRED_RPROVIDER_sshd="dropbear".

Signed-off-by: Jate Sujjavanich <jatedev@gmail.com>