providers.py: Add package name to PREFERRED_PROVIDER search
[Yocto Bug #5044] - https://bugzilla.yoctoproject.org/show_bug.cgi?id=5044
[Yocto Bug #6149] - https://bugzilla.yoctoproject.org/show_bug.cgi?id=6149

Add item (package name) to list that is searched for possible
PREFERRED_PROVIDER's.

The WARNING from bitbake suggested specifying a PREFERRED_PROVIDER for the
package name, but setting it did nothing because the filterRunTimeProviders
searched only for the recipe names.

For the example case in bug #6149, bitbake looks for PREFERRED_PROVIDER_openssh
and PREFERRED_provider_dropbear. This change allows
PREFERRED_PROVIDER_sshd="dropbear".

Signed-off-by: Jate Sujjavanich <jatedev@gmail.com>