image.bbclass/rootfs.py: use PREFERRED_RPROVIDER to select package
[Yocto Bug #6149] - https://bugzilla.yoctoproject.org/show_bug.cgi?id=6149

Add rootfs_process_preferred_rproviders to image.bbclass. In PACKAGE_INSTALL,
this substitutes the requested package with the provider. Change regex in
RpmRootfs._log_check_error to allow log message about PREFERRED_RPROVIDER.

For example, an image recipe might include sshd in IMAGE_INSTALL. The recipe
dropbear sets RPROVIDES_dropbear = "sshd", and image.bbclass does not know
about this. It causes an error during do_rootfs. This change substitues
sshd with dropbear in the PACKAGE_INSTALL variable which is used to determine
which packages to install.

Signed-off-by: Jate Sujjavanich <jatedev@gmail.com>